Anh hao

+++ QuanLyChuoiCuaHangTienLoi_Ministop/ET/ET_KhuyenMai.cs
@@ -10,11 +10,11 @@ namespace ET
     {
       public string MaKhuyenMai { get; set; }
       public string TenKhuyenMai { get; set; }
-      public double MucGiamGia { get; set; }
+      public decimal MucGiamGia { get; set; }
       public DateTime NgayBatDau { get; set; }
       public DateTime NgayKetThuc { get; set; }
       public string LoaiApDung { get; set; }
-      public double DieuKienApDung { get; set; }
+      public decimal DieuKienApDung { get; set; }
       public int TrangThai { get; set; }
     }
 }
